refactor documentation settings
