+++ src/main/scala/nl/biopet/sbtbiopet/BiopetPlugin.scala
@@ -33,19 +33,8 @@ import com.lucidchart.sbt.scalafmt.ScalafmtSbtPlugin.autoImport.Sbt
 import com.typesafe.sbt.SbtGit.git
 import com.typesafe.sbt.SbtPgp.autoImport.useGpg
 import com.typesafe.sbt.sbtghpages.GhpagesPlugin
-import com.typesafe.sbt.sbtghpages.GhpagesPlugin.autoImport.{
-  ghpagesCleanSite,
-  ghpagesPushSite,
-  ghpagesRepository
-}
-import com.typesafe.sbt.site.SitePlugin.autoImport.{
-  makeSite,
-  siteDirectory,
-  siteSubdirName
-}
-import com.typesafe.sbt.site.SiteScaladocPlugin.autoImport.SiteScaladoc
+import com.typesafe.sbt.sbtghpages.GhpagesPlugin.autoImport.ghpagesRepository
 import com.typesafe.sbt.site.laika.LaikaSitePlugin
-import com.typesafe.sbt.site.laika.LaikaSitePlugin.autoImport.LaikaSite
 import com.typesafe.sbt.site.{SitePlugin, SiteScaladocPlugin}
 import de.heikoseeberger.sbtheader.HeaderPlugin.autoImport.{
   HeaderCommentStyle,
@@ -54,11 +43,11 @@ import de.heikoseeberger.sbtheader.HeaderPlugin.autoImport.{
   headerMappings
 }
 import de.heikoseeberger.sbtheader.{FileType, HeaderPlugin}
-import laika.sbt.LaikaPlugin.autoImport.{Laika, laikaRawContent}
 import nl.biopet.bioconda.BiocondaPlugin
 import nl.biopet.bioconda.BiocondaPlugin.autoImport._
-import nl.biopet.sbtbiopet.BiopetReleaseSteps._
-import nl.biopet.utils.Documentation.{htmlRedirector, markdownExtractChapter}
+import nl.biopet.sbtbiopet.BiopetDocumentationSettings._
+import nl.biopet.sbtbiopet.BiopetReleaseSettings._
+import nl.biopet.utils.Documentation.markdownExtractChapter
 import ohnosequences.sbt.SbtGithubReleasePlugin.autoImport._
 import ohnosequences.sbt.{GithubRelease, SbtGithubReleasePlugin}
 import sbt.Keys._
@@ -163,6 +152,7 @@ object BiopetPlugin extends AutoPlugin {
       }
     )
   }
+
   /*
    * Contains al settings related to the license header
    */
@@ -189,6 +179,7 @@ object BiopetPlugin extends AutoPlugin {
         .dependsOn(headerCheck in Test)
         .value
     )
+
   /*
    * A sequence of settings containing information such as homepage, licences and git related information.
    */
@@ -243,43 +234,6 @@ object BiopetPlugin extends AutoPlugin {
     releaseProcess := biopetReleaseProcess.value
   )
 
-  /*
-   * A sequence of settings related to documentation.
-   * This includes all the settings for
-   *  - LAIKA
-   *  - Ghpagesplugin
-   *  - Our custom documentation generation code
-   *  - SBT-site
-   */
-  protected def biopetDocumentationSettings: Seq[Setting[_]] = Seq(
-    biopetDocsDir := file(
-      s"%s${File.separator}markdown".format(target.value.toString)),
-    biopetReadmePath := file("README.md").getAbsoluteFile,
-    sourceDirectory in LaikaSite := biopetDocsDir.value,
-    sourceDirectories in Laika := Seq((sourceDirectory in LaikaSite).value),
-    siteDirectory in Laika := file(
-      target.value.toString + s"${File.separator}site"),
-    ghpagesRepository := file(target.value.toString + s"${File.separator}gh"),
-    siteSubdirName in SiteScaladoc := {
-      if (biopetIsTool.value) {
-        if (isSnapshot.value) {
-          s"develop${File.separator}api"
-        } else s"${version.value}${File.separator}api"
-      } else {
-        if (isSnapshot.value) {
-          "develop"
-        } else s"${version.value}"
-      }
-    },
-    laikaRawContent in LaikaSite := true, //Laika use raw HTML content in markdown.
-    includeFilter in ghpagesCleanSite := biopetCleanSiteFilter.value,
-    biopetGenerateDocs := biopetGenerateDocsFunction().value,
-    biopetGenerateReadme := biopetGenerateReadmeFunction().value,
-    makeSite := (makeSite triggeredBy biopetGenerateDocs).value,
-    makeSite := (makeSite dependsOn biopetGenerateDocs).value,
-    ghpagesPushSite := (ghpagesPushSite dependsOn makeSite).value
-  )
-
   protected def biopetScalafmtSettings: Seq[Setting[_]] = Seq(
     scalafmtOnCompile := true, // make sure scalafmt is run regularly during development
     // make sure scalafmt command reformats everything
@@ -315,7 +269,10 @@ object BiopetPlugin extends AutoPlugin {
               // Remove whitespace from beginning and end of string.
               .trim)
             }
-        } else Def.task { Some("") }
+        } else
+          Def.task {
+            Some("")
+          }
       }
       .dependsOn(biopetGenerateReadme)
       .value,
@@ -328,17 +285,23 @@ object BiopetPlugin extends AutoPlugin {
               markdownExtractChapter(readme, name.value, includeHeader = false)
             // Assuming the first sentence ends with .
             description.split("\\.").headOption match {
-              case Some(s) => { s + "." }.replace("\n", " ").trim
+              case Some(s) => {
+                s + "."
+              }.replace("\n", " ").trim
               case _ =>
                 s"This summary for ${(name in Bioconda).value} was automatically generated."
             }
 
           }
-        } else Def.task { "" }
+        } else
+          Def.task {
+            ""
+          }
       }
       .dependsOn(biopetGenerateReadme)
       .value
   )
+
   /*
    * The merge strategy that is used in biopet projects
    */
@@ -383,13 +346,6 @@ object BiopetPlugin extends AutoPlugin {
       else
         Some(Opts.resolver.sonatypeStaging)
     }
-
-  /*
-   * The filter that is used by the ghpages plugin.
-   * All files in this filter will be removed.
-   * This allows the updating of documentation for a specific version
-   * All other versions will not be touched.
-   */
   protected def biopetCleanSiteFilter: Def.Initialize[FileFilter] =
     Def.setting {
       new FileFilter {
@@ -410,86 +366,4 @@ object BiopetPlugin extends AutoPlugin {
       }
     }
 
-  /*
-   * Accesses the tools main method to generate documentation using our custom built-in documentation function
-   */
-  protected def biopetGenerateDocsFunction(): Def.Initialize[Task[Unit]] =
-    Def.taskDyn {
-      if (biopetIsTool.value) {
-        Def
-          .task[Unit] {
-            val r = (runner in Compile).value
-            val classPath = (fullClasspath in Compile).value
-
-            val streamsLogValue = streams.value.log
-
-            val args = Seq("--generateDocs",
-                           s"outputDir=${biopetDocsDir.value.toString}," +
-                             s"version=${version.value}," +
-                             s"release=${!isSnapshot.value}",
-                           version.value)
-
-            val mainClassString = (mainClass in assembly).value match {
-              case Some(x) => x
-              case _ =>
-                throw new IllegalStateException(
-                  "Mainclass should be defined for a tool.")
-            }
-            import Attributed.data
-            r.run(
-              mainClassString,
-              data(classPath),
-              args,
-              streamsLogValue
-            )
-
-          }
-          .dependsOn(compile in Compile)
-      } else
-        Def.task[Unit] {
-          biopetDocsDir.value.mkdirs()
-          if (!isSnapshot.value) {
-            val htmlRedirectFile: sbt.File = biopetDocsDir.value / "index.html"
-            htmlRedirector(
-              outputFile = htmlRedirectFile,
-              link = s"${version.value}${File.separator}index.html",
-              title = "API documentation",
-              redirectText = "Go to the API documentation")
-          }
-        }
-    }
-
-  /*
-   * Accesses the tools main method to generate a README using our custom built-in documentation function
-   */
-  protected def biopetGenerateReadmeFunction(): Def.Initialize[Task[Unit]] =
-    Def.taskDyn {
-      if (biopetIsTool.value) {
-        Def
-          .task[Unit] {
-            val r = (runner in Compile).value
-            val classPath = (fullClasspath in Compile).value
-
-            val args = Seq("--generateReadme", biopetReadmePath.value.toString)
-
-            val streamsLogValue = streams.value.log
-
-            val mainClassString = (mainClass in assembly).value match {
-              case Some(x) => x
-              case _ =>
-                throw new IllegalStateException(
-                  "Mainclass should be defined for a tool.")
-            }
-            import Attributed.data
-            r.run(
-              mainClassString,
-              data(classPath),
-              args,
-              streamsLogValue
-            )
-
-          }
-          .dependsOn(compile in Compile)
-      } else Def.task[Unit] {}
-    }
 }
